⚡️ feat: add cleanup job and refactor the code
